Automatic Attendence using Face Recognition/requirements.txt
absl-py==0.12.0
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.0.3
astroid==2.4.0
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.5
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
binaryornot==0.4.4
black==20.8b1
bleach==3.3.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
cookiecutter==1.7.2
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
docutils==0.17
entrypoints==0.3
flake8==3.8.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
imagesize==1.2.0
importlib-metadata==3.10.0
imutils==0.5.4
inflection==0.5.1
intervaltree==3.1.0
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-vggface==0.6
keyring==23.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mtcnn==0.1.0
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.3.0
numpy==1.19.5
numpydoc==1.1.0
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
poyo==0.5.0
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.15.7
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==6.0.0
pyflakes==2.2.0
Pygments==2.8.1
pylint==2.5.0
pyls-black==0.4.6
pyls-spyder==0.3.2
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.1
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
PySimpleGUI==4.38.0
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-slugify==4.0.1
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==22.0.3
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.2
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
rope-py3k==0.9.4.post1
rsa==4.7.2
scikit-learn==0.24.1
scipy @ file:///C:/ci/scipy_1616703433439/work
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
Sphinx==3.5.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==5.0.0
spyder-kernels==2.0.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.1.0
three-merge==0.1.1
tinycss==0.4
toml==0.10.2
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.4
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
yapf==0.31.0
zipp==3.4.1
